Update paths to files moved lib
