+++ challenges/branches_arent_just_for_birds.html
@@ -225,7 +225,7 @@ GitHub Guide: <a href="http://guides.github.com/overviews/flow/" target="_blank"
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'branches_arent_just_for_birds'
       challengeStatus.completed(currentChallenge)
     </script>
