+++ challenges/commit_to_it.html
@@ -180,7 +180,7 @@
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'commit_to_it'
       challengeStatus.completed(currentChallenge)
     </script>
