+++ challenges/forks_and_clones.html
@@ -193,7 +193,7 @@ Now clone:
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'forks_and_clones'
       challengeStatus.completed(currentChallenge)
     </script>
