+++ challenges/get_git.html
@@ -163,7 +163,7 @@ type the <code>$</code> in though, only type what comes after. For instance, to 
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'get_git'
       challengeStatus.completed(currentChallenge)
     </script>
