+++ challenges/githubbin.html
@@ -141,7 +141,7 @@ which will be needed in order to verify the upcoming challenges. Save it exactly
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'githubbin'
       challengeStatus.completed(currentChallenge)
     </script>
