+++ challenges/its_a_small_world.html
@@ -128,7 +128,7 @@
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'its_a_small_world'
       challengeStatus.completed(currentChallenge)
     </script>
