+++ challenges/merge_tada.html
@@ -184,7 +184,7 @@
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'merge_tada'
       challengeStatus.completed(currentChallenge)
     </script>
