+++ challenges/pull_never_out_of_date.html
@@ -163,7 +163,7 @@
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'pull_never_out_of_date'
       challengeStatus.completed(currentChallenge)
     </script>
