+++ challenges/remote_control.html
@@ -230,7 +230,7 @@ want to tell Git the location of the remote version on GitHub's servers. You can
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'remote_control'
       challengeStatus.completed(currentChallenge)
     </script>
