+++ challenges/repository.html
@@ -183,7 +183,7 @@
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'repository'
       challengeStatus.completed(currentChallenge)
     </script>
