+++ challenges/requesting_you_pull_please.html
@@ -161,7 +161,7 @@ request.</p>
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = 'requesting_you_pull_please'
       challengeStatus.completed(currentChallenge)
     </script>
