+++ index.html
@@ -8,7 +8,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <link rel="icon" href="assets/imgs/favicon.png" type="image/x-icon">
         <link rel="stylesheet" href="assets/css/style.css">
-        <script src="./challenges-completed.js"></script>
+        <script src="./lib/challenges-completed.js"></script>
         <script src="./lib/handle-external-links.js"></script>
     </head>
   <body class="tableofcontents">
