+++ layouts/challenge.hbs
@@ -39,7 +39,7 @@
     </script>
 
     <script>
-      var challengeStatus = require('../challenge-completed.js')
+      var challengeStatus = require('../lib/challenge-completed.js')
       var currentChallenge = '{{{shortname}}}'
       challengeStatus.completed(currentChallenge)
     </script>
