+++ lib/buildpages.js
@@ -12,7 +12,6 @@ var thefiles = []
 var lang = process.argv[2]
 var rawFiles = path.join(__dirname, (lang ? '../raw-content-' + lang + '/' : '../raw-content/'))
 var builtContent = path.join(__dirname, (lang ? '../challenges-' + lang + '/' : '../challenges/'))
-console.log(rawFiles, builtContent)
 // I can probably use glob better to avoid
 // finding the right files within the files
 glob('*.html', {cwd: rawFiles}, function (err, files) {
@@ -23,7 +22,6 @@ glob('*.html', {cwd: rawFiles}, function (err, files) {
 
 function buildPage (files) {
   files.forEach(function (file) {
-    console.log(file)
     // shouldn't have to do this if my
     // mapping were correct
     if (!file) return
