+++ lib/challenge-completed.js
