+++ lib/challenges-completed.js
@@ -1,7 +1,7 @@
 var ipc = require('ipc')
 var fs = require('fs')
 
-var userData = require('./user-data.js')
+var userData = require('./lib/user-data.js')
 
 document.addEventListener('DOMContentLoaded', function (event) {
   var data = userData.getData()
