+++ verify/get_git.js
@@ -1,7 +1,7 @@
 var exec = require('child_process').exec
 
 var helper = require('../verify/helpers.js')
-var userData = require('../user-data.js')
+var userData = require('../lib/user-data.js')
 
 var addToList = helper.addtoList
 var markChallengeCompleted = helper.markChallengeCompleted
